+++ examples/receipt.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 require_once __DIR__ . '/../vendor/autoload.php';
 
-$engine = (new \Folio\Pdf\Template\TemplateEngine())->enableFolio2Syntax(__DIR__ . '/templates');
+$engine = new \Folio\Pdf\Template\TemplateEngine();
 
 $pdf = $engine->renderFile(__DIR__ . '/templates/receipt.folio', [
     'store' => 'Byte Cafe',